Rename Storage to StorageReader (#467)
