+++ src/StorageReader/Entry.ts
