+++ src/StorageReader/Router.test.ts
@@ -7,8 +7,9 @@ import { ClaimController } from './ClaimController'
 import { ClaimControllerConfiguration } from './ClaimControllerConfiguration'
 import { IPFS } from './IPFS'
 import { IPFSConfiguration } from './IPFSConfiguration'
+import { Router } from './Router'
 
-describe('Storage ClaimController', async (should: any) => {
+describe('StorageReader Router', async (should: any) => {
   const { assert } = should('')
 
   const host = 'http://localhost'
@@ -24,19 +25,21 @@ describe('Storage ClaimController', async (should: any) => {
     downloadMaxAttempts: 1,
   }
 
+  const claimController = new ClaimController(
+    Pino(),
+    new Db('poet', server),
+    new Messaging(),
+    new IPFS(IPFSConfiguration),
+    claimControllerConfiguration
+  )
+
   {
-    const claimController = new ClaimController(
-      Pino(),
-      new Db('poet', server),
-      new Messaging(),
-      new IPFS(IPFSConfiguration),
-      claimControllerConfiguration
-    )
+    const router = new Router(Pino(), new Messaging(), claimController)
 
     assert({
-      given: 'the new instance of ClaimController',
-      should: 'be an instance of ClaimController',
-      actual: claimController instanceof ClaimController,
+      given: 'the new instance of Router',
+      should: 'be an instance of Router',
+      actual: router instanceof Router,
       expected: true,
     })
   }
