+++ src/StorageReader/Service.test.ts
@@ -7,9 +7,10 @@ import { ClaimController } from './ClaimController'
 import { ClaimControllerConfiguration } from './ClaimControllerConfiguration'
 import { IPFS } from './IPFS'
 import { IPFSConfiguration } from './IPFSConfiguration'
-import { Router } from './Router'
+import { Service } from './Service'
+import { ServiceConfiguration } from './ServiceConfiguration'
 
-describe('Storage Router', async (should: any) => {
+describe('StorageReader Router', async (should: any) => {
   const { assert } = should('')
 
   const host = 'http://localhost'
@@ -33,13 +34,17 @@ describe('Storage Router', async (should: any) => {
     claimControllerConfiguration
   )
 
+  const serviceConfiguration: ServiceConfiguration = {
+    downloadIntervalInSeconds: 1,
+  }
+
   {
-    const router = new Router(Pino(), new Messaging(), claimController)
+    const service = new Service(Pino(), claimController, serviceConfiguration)
 
     assert({
-      given: 'the new instance of Router',
-      should: 'be an instance of Router',
-      actual: router instanceof Router,
+      given: 'the new instance of Service',
+      should: 'be an instance of Service',
+      actual: service instanceof Service,
       expected: true,
     })
   }
