+++ tests/unit/index.ts
@@ -15,10 +15,10 @@ import '../../src/Helpers/Time.test.ts'
 import '../../src/Helpers/camelCaseToScreamingSnakeCase.test'
 import '../../src/Helpers/isTruthy.test'
 import '../../src/Interfaces.test.ts'
-import '../../src/Storage/ClaimController.test.ts'
-import '../../src/Storage/IPFS.test.ts'
-import '../../src/Storage/Router.test.ts'
-import '../../src/Storage/Service.test.ts'
+import '../../src/StorageReader/ClaimController.test.ts'
+import '../../src/StorageReader/IPFS.test.ts'
+import '../../src/StorageReader/Router.test.ts'
+import '../../src/StorageReader/Service.test.ts'
 import '../../src/StorageWriter/ClaimController.test.ts'
 import '../../src/StorageWriter/IPFS.test.ts'
 import '../../src/StorageWriter/Router.test.ts'
